requirements/requirements_dev.txt
-r requirements_test.txt

# Debug
django-debug-panel

django-rainbowtests
django_extensions

model_mommy
